Create cpu.py
File with the code of the core part of the cpu: rules to process the instructions via the instructions registers, rules to execute individual instructions based on the MIPS assembly language, management of the programm counter, of the cache. 
At the end of the file, call to populate the memory plus a call for executing the instructions from the instructions file.